requirements.txt
pvporcupine
sounddevice
vosk
pyttsx3
requests
python-dotenv
